requirements.txt
Flask~=2.0.1
flask-sqlalchemy
flask-bootstrap
flask-nav
Flask-FontAwesome
pymysql
python-dotenv
Flask-WTF